\name{CarbonExample3Data}
\alias{CarbonExample3Data}
\docType{data}
\title{
Data from Carbon, C. C. (2013). BiDimRegression: Bidimensional regression modeling using R. Journal of Statistical Software
}
\description{
Example 3 focuses on demonstrating how good a cognitive map recalculated from averaged cognitive distance data fits with a related real map
}
\usage{data(CarbonExample3Data)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{depV1}}{a numeric vector}
    \item{\code{depV2}}{a numeric vector}
    \item{\code{indepV1}}{a numeric vector}
    \item{\code{indepV2}}{a numeric vector}
  }
}
\examples{
data(CarbonExample3Data)
## maybe str(CarbonExample3Data) ; plot(CarbonExample3Data) ...
}
\keyword{datasets}
