\name{FriedmanKohlerData2}
\alias{FriedmanKohlerData2}
\docType{data}
\title{
data from Friedman, A., & Kohler, B. (2003). Bidimensional regression: Assessing the configural similarity and accuracy of cognitive maps and other   two-dimensional data sets. Psychological Methods, 8(4), 468-491.
}
\description{
data from Friedman, A., & Kohler, B. (2003). Bidimensional regression: Assessing the configural similarity and accuracy of cognitive maps and other   two-dimensional data sets. Psychological Methods, 8(4), 468-491.
}
\usage{data(FriedmanKohlerData2)}
\format{
  A data frame with 4 observations on the following 4 variables.
  \describe{
    \item{\code{depV1}}{a numeric vector}
    \item{\code{depV2}}{a numeric vector}
    \item{\code{indepV1}}{a numeric vector}
    \item{\code{indepV2}}{a numeric vector}
  }
}
\examples{
data(FriedmanKohlerData2)
## maybe str(FriedmanKohlerData2) ; plot(FriedmanKohlerData2) ...
}
\keyword{datasets}
