% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidimensional.R
\name{lm2fit}
\alias{lm2fit}
\title{Fits the specified model and computes stats}
\usage{
lm2fit(data, transformation)
}
\arguments{
\item{data}{the preprocessed data frame from \code{\link{lm2}} function,
so that the first two columns are the dependent variables and the other
two are independent variables}

\item{transformation}{the transformation to be used, either \code{'euclidean'} or \code{'affine'}.}
}
\value{
returns an object of class "lm2", see \code{\link{lm2}}
for the description.
}
\description{
Calls a specific transformation model function and then computes statistics
that is common across all transformations.
This function should not be called directly, please use \code{\link{lm2}}.
}
\keyword{internal}
