% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidimensional.R
\name{lm2projective}
\alias{lm2projective}
\title{Computes model for the projective transformation}
\usage{
lm2projective(data)
}
\arguments{
\item{data}{the preprocessed data frame from \code{\link{lm2}} function,
so that the first two columns are the dependent variables and the other
two are independent variables}
}
\value{
object with transformation specific data to be supplemented with further stats
}
\description{
Computes model for the projective transformation
}
\keyword{internal}
