\name{NakayaData}
\alias{NakayaData}
\docType{data}
\title{
Data from Nakaya, T. (1997). Statistical inferences in bidimensional regression models. Geographical Analysis, 29(2), 169-186.
}
\description{
Data from Nakaya, T. (1997). Statistical inferences in bidimensional regression models. Geographical Analysis, 29(2), 169-186.
}
\usage{data(NakayaData)}
\format{
  A data frame with 19 observations on the following 4 variables.
  \describe{
    \item{\code{depV1}}{a numeric vector}
    \item{\code{depV2}}{a numeric vector}
    \item{\code{indepV1}}{a numeric vector}
    \item{\code{indepV2}}{a numeric vector}
  }
}
\examples{
data(NakayaData)
## maybe str(NakayaData) ; plot(NakayaData) ...
}
\keyword{datasets}
