\name{print.summary.BiDimRegression}
\alias{print.summary.BiDimRegression}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function to print the summary of the results of BiDimRegression
}
\description{
function to print the summary of the results of BiDimRegression
}
\references{
Carbon, C. C. (2013). BiDimRegression: Bidimensional regression modeling using R. Journal of Statistical Software.
}
\author{
Claus-Christian Carbon
}

