% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidimensional.R
\name{summary.lm2}
\alias{summary.lm2}
\title{Makes a lightweight summary lm2 object}
\usage{
\method{summary}{lm2}(object, ...)
}
\arguments{
\item{object}{an object of class "lm2", see \code{\link{lm2}}}
}
\description{
Drops heavy bits, like the data frame with predicted values or the lm object.
However, the print tells more! :)
}
\keyword{internal}
