\name{BiDimRegression}

\alias{BiDimRegression}

\title{function to calculate the bidimensional regression between two 2D configurations}

\description{An S3 class with a method for calculates the bidimensional regression between two 2D configurations

following the approach by Tobler (1965).}

\references{*** The routine is described in detail in:\\
Carbon, C. C. (2013). BiDimRegression: Bidimensional Regression Modeling Using R. Journal of Statistical Software,\\
Code Snippets, 52(1), 1-11 (URL http://www.jstatsoft.org/v52/c01/)\\
\\
*** The routine is based on the following literature:\\
1) Tobler, W. [R.](1965). Computation of the corresponding of geographical patterns. Papers of the Regional Science Association, 15, 131-139.\\
2) Tobler, W. R. (1966). Medieval distortions: Projections of ancient maps. Annals of the Association of American Geographers, 56(2), 351-360.\\
3) Tobler, W. R. (1994). Bidimensional regression. Geographical Analysis, 26(3), 187-212.\\
4) Friedman, A., & Kohler, B. (2003). Bidimensional regression: Assessing the configural similarity and accuracy of cognitive maps and other two-dimensional data sets. Psychological Methods, 8(4), 468-491.\\
5) Nakaya, T. (1997). Statistical inferences in bidimensional regression models. Geographical Analysis, 29(2), 169-186.\\
6) Waterman, S., & Gordon, D. (1984). A quantitative-comparative approach to analysis of distortion in mental maps. Professional Geographer\\
, 36(3), 326-337.}

\author{Claus-Christian Carbon}

\note{The author is grateful to Waldo R. Tobler, now Professor Emeritus at the Department of Geography, University of California, Santa Barbara, for providing his original publications and his helpful correspondence and to Dan Montello for calling the author's attention to Tobler's work many years ago. I would like to thank Alinda Friedman, Gregory Francis, Jan de Leeuw, Achim Zeileis, two anonymous reviewers and Arne Terkowski and Chris Nappert for valuable comments on an earlier version of this paper, and Andrea Lyman and Vera M. Hesslinger for proofreading the manuscript. Last but not least, I am very indepted to Tomoki Nakaya, who has developed the original inference statistics of the overall models and the referring parameters and who helped me with reanalyzing these statistics as well as with ensuring the reliability of the used methods. Thank you!}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ regression }
\keyword{ bidimensional }
\keyword{ Pearson }
\keyword{ correlation }
\keyword{ AIC }
\keyword{ similarity }
\keyword{ comparison }

