\name{CarbonExample1Data}
\alias{CarbonExample1Data}
\docType{data}
\title{
Data from Carbon, C. C. (2013). BiDimRegression: Bidimensional Regression Modeling Using R. \\
Journal of Statistical Software, Code Snippets, 52(1), 1-11 (URL http://www.jstatsoft.org/v52/c01/)\\
}
\description{
Example 1 from the domain of aesthetics to show how the method can be utilized for assessing the similarity of two portrayed persons, actually the Mona Lisa in the world famous Louvre version and the only recently re-discovered Prado version
}
\usage{data(CarbonExample1Data)}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{depV1}}{a numeric vector}
    \item{\code{depV2}}{a numeric vector}
    \item{\code{indepV1}}{a numeric vector}
    \item{\code{indepV2}}{a numeric vector}
  }
}
\examples{
data(CarbonExample1Data)
## maybe str(CarbonExample1Data) ; plot(CarbonExample1Data) ...
}
\keyword{datasets}
