\name{CarbonExample2Data}
\alias{CarbonExample2Data}
\docType{data}
\title{
data from Carbon, C. C. (2013). BiDimRegression: Bidimensional regression modeling using R. Journal of Statistical Software
}
\description{
 Example 2 originates from the area of geography and inspects the accuracy of different maps of the city of Paris which were created over the last 350 years as compared to a recent map}
\usage{data(CarbonExample2Data)}
\format{
  A data frame with 13 observations on the following 4 variables.
  \describe{
    \item{\code{depV1}}{a numeric vector}
    \item{\code{depV2}}{a numeric vector}
    \item{\code{indepV1}}{a numeric vector}
    \item{\code{indepV2}}{a numeric vector}
  }
}
\examples{
data(CarbonExample2Data)
## maybe str(CarbonExample2Data) ; plot(CarbonExample2Data) ...
}
\keyword{datasets}
