\name{summary.BiDimRegression}
\alias{summary.BiDimRegression}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
function to provide a summary of the results of BiDimRegression
}
\description{
function to provide a summary of the results of BiDimRegression
}
\references{
Data from Carbon, C. C. (2013). BiDimRegression: Bidimensional Regression Modeling Using R. \\
Journal of Statistical Software, Code Snippets, 52(1), 1-11 (URL http://www.jstatsoft.org/v52/c01/)\\
}
\author{
Claus-Christian Carbon
}

